\name{BMRankMerging}
\alias{BMRankMerging}
\title{Merging two or more selected ranked lists into a new one ranked list
}
\description{Implements a majority voting system, use the Borda Merging Method to merging two or more ranked lists into single one list. This method is used in function \link{krubor}.
}
\usage{
BMRankMerging(rankings)
}
\arguments{
  \item{rankings}{a matrix or a data.frame, which must be numeric.
}
}
\references{Lin S. (2010) \emph{Space oriented rank-based data integration}
}


