% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSignature-package.R
\docType{data}
\name{logfcData}
\alias{logfcData}
\title{LogFC sample data}
\format{data.frame}
\description{
LogFC sample data, contining logFC values, have 5 samples with 4356 genes.
}
\examples{

data(exampleSet)

}
\author{
Zhilong Jia
}
\keyword{data}
\keyword{datasets}

