% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExpressionSet.R
\name{makeExpressionSet}
\alias{makeExpressionSet}
\title{Convert data.frame into ExpressionSet Object.}
\usage{
makeExpressionSet(dat)
}
\arguments{
\item{dat}{a data.frame with gene in row, sample in column}
}
\value{
an ExpressionSet object
}
\description{
Convert data.frame into ExpressionSet Object.
}
\examples{
dat <- data.frame(d1=1:6, d2=6:1)
datExp <- makeExpressionSet(dat)

}

