% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreX.R
\name{ScoreX}
\alias{ScoreX}
\title{ScoreX}
\usage{
ScoreX(obj, metric = "XSum", SignatureLength = 500, scale = TRUE, ...)
}
\arguments{
\item{obj}{Either a numeric (logFC) matrix, a data.frame, or an ExpressionSet object.}

\item{metric}{simlarity metric. Such as XSum, XCos.}

\item{SignatureLength}{the length of "gene signature".}

\item{scale}{scale or not.}

\item{...}{other parameters}
}
\description{
ScoreX: The eXtreme simlarity score
}
\examples{
data(exampleSet)
ScoreXSum <- ScoreX(logfcData)
ScoreXCos <- ScoreX(logfcData, metric="XCos")
}
\author{
Zhilong Jia
}
\references{
Cheng, Jie, and Lun Yang. "Comparing gene expression similarity metrics for 
connectivity map." Bioinformatics and Biomedicine (BIBM), 2013 IEEE 
International Conference on. IEEE, 2013.
}

