\name{exampleSet}
\alias{exampleSet}
\docType{data}
\title{sample data, a subset of the C-MAP 
}
\description{sample data, a subset of the C-MAP as , which is a collection of 50 genome-wide transcriptional 
expression data from cultured human cells treated with 15 different small molecules.
}
\usage{data(exampleSet)}
\format{
  A ExpressionSet: assay data represents the 50 genome-wide transcriptional expression data, phenotypic data describes
15 different small molecules corresponds to the expression data (assay data).
}
\references{
  \url{http://www.sciencemag.org/content/313/5795/1929.short}
  Lamb et al., The Connectivity Map: Using Gene-Expression 
  Signatures to Connect Small Molecules, Genes, and Disease, science 2006
}
\examples{
data(exampleSet)
}
\keyword{data}
\keyword{datasets}
