% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSignature.R
\docType{package}
\name{GeneExpressionSignature}
\alias{GeneExpressionSignature}
\title{Gene Expression Signature based Similarity Metric}
\description{
This package gives the implementations of the gene expression signature
and its distance to each. Gene expression signature is represented as a list
of genes whose expression is correlated with a biological state of interest.
And its distance is defined using a nonparametric, rank-based pattern-
matching strategy based on the Kolmogorov-Smirnov statistic. Gene expression
signature and its distance can be used to detect similarities among the
signatures of drugs, diseases, and biological states of interest.
}
